Extracted the search api to a new project
