--- api/src/main/java/com/github/tsar_industries/slashgif/api/image_search/ImageSearchApi.java
@@ -1,8 +1,0 @@
-package com.github.tsar_industries.slashgif.api.image_search;
-
-import java.util.Optional;
-
-public interface ImageSearchApi
-{
-	Optional<Image> getImage(String query);
-}
