+++ api/src/main/java/com/github/tsar_industries/slashgif/api/model/Image.java
@@ -1,0 +1,5 @@
+package com.github.tsar_industries.slashgif.api.model;
+
+public class Image
+{
+}
