+++ pom.xml
@@ -48,6 +48,11 @@
 				<artifactId>google-image-search</artifactId>
 				<version>1.0-SNAPSHOT</version>
 			</dependency>
+			<dependency>
+				<groupId>com.github.tsar-industries.slashgif</groupId>
+				<artifactId>search-provder-api</artifactId>
+				<version>1.0-SNAPSHOT</version>
+			</dependency>
 
 			<dependency>
 				<groupId>io.dropwizard</groupId>
