+++ search-providers/google-image-search/pom.xml
@@ -14,6 +14,10 @@
 	<dependencies>
 		<dependency>
 			<groupId>com.github.tsar-industries.slashgif</groupId>
+			<artifactId>search-provder-api</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>com.github.tsar-industries.slashgif</groupId>
 			<artifactId>api</artifactId>
 		</dependency>
