+++ search-providers/google-image-search/src/main/java/com/github/tsar_industries/slashgif/search_providers/google_image_search/GoogleSearchProvider.java
@@ -1,13 +1,13 @@
-package com.github.tsar_industries.slashgif.google;
+package com.github.tsar_industries.slashgif.search_providers.google_image_search;
 
-import com.github.tsar_industries.slashgif.api.image_search.Image;
-import com.github.tsar_industries.slashgif.api.image_search.ImageSearchApi;
+import com.github.tsar_industries.slashgif.api.model.Image;
+import com.github.tsar_industries.slashgif.search_providers.api.SearchProviderApi;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Optional;
 
-class GoogleImageSearch implements ImageSearchApi
+class GoogleSearchProvider implements SearchProviderApi
 {
 	private final Logger logger = LoggerFactory.getLogger(getClass());
