+++ search-providers/search-provider-api/pom.xml
@@ -3,16 +3,17 @@
 		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<parent>
-		<artifactId>slashgif-parent</artifactId>
+		<artifactId>search-providers</artifactId>
 		<groupId>com.github.tsar-industries.slashgif</groupId>
 		<version>1.0-SNAPSHOT</version>
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
-	<packaging>pom</packaging>
 
-	<artifactId>search-providers</artifactId>
-	<modules>
-		<module>google-image-search</module>
-		<module>search-provider-api</module>
-	</modules>
+	<artifactId>search-provder-api</artifactId>
+	<dependencies>
+		<dependency>
+			<groupId>com.github.tsar-industries.slashgif</groupId>
+			<artifactId>api</artifactId>
+		</dependency>
+	</dependencies>
 </project>
