+++ search-providers/search-provider-api/src/main/java/com/github/tsar_industries/slashgif/search_providers/api/SearchProviderApi.java
@@ -1,11 +1,10 @@
-package com.github.tsar_industries.slashgif.api.service;
-
+package com.github.tsar_industries.slashgif.search_providers.api;
 
 import com.github.tsar_industries.slashgif.api.model.Image;
 
 import java.util.Optional;
 
-public interface ImageSearchService
+public interface SearchProviderApi
 {
 	Optional<Image> getImage(String query);
 }
