+++ service/src/main/java/com/github/tsar_industries/slashgif/service/PlaceHolderImageSearchService.java
@@ -1,9 +1,9 @@
 package com.github.tsar_industries.slashgif.service;
 
 
-import com.github.tsar_industries.slashgif.api.image_search.Image;
-import com.github.tsar_industries.slashgif.api.image_search.ImageSearchApi;
+import com.github.tsar_industries.slashgif.api.model.Image;
 import com.github.tsar_industries.slashgif.api.service.ImageSearchService;
+import com.github.tsar_industries.slashgif.search_providers.api.SearchProviderApi;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -14,18 +14,18 @@ class PlaceHolderImageSearchService implements ImageSearchService
 {
 	private final Logger logger = LoggerFactory.getLogger(getClass());
 
-	private final ImageSearchApi imageSearcher;
+	private final SearchProviderApi searchProvider;
 
 	@Inject
-	PlaceHolderImageSearchService(ImageSearchApi imageSearcher)
+	PlaceHolderImageSearchService(SearchProviderApi searchProvider)
 	{
-		this.imageSearcher = imageSearcher;
+		this.searchProvider = searchProvider;
 	}
 
 	@Override
 	public Optional<Image> getImage(String query)
 	{
 		logger.info("Searching for images for the query \"{}\"", query);
-		return imageSearcher.getImage(query);
+		return searchProvider.getImage(query);
 	}
 }
